+++ aether-common/src/main/java/com/xiaobai1226/aether/common/util/ImageUtils.java
@@ -3,14 +3,7 @@ package com.xiaobai1226.aether.common.util;
 import cn.hutool.core.io.FileUtil;
 import cn.hutool.core.io.IoUtil;
 import cn.hutool.core.util.RuntimeUtil;
-import com.github.kokorin.jaffree.ffmpeg.FFmpeg;
-import com.github.kokorin.jaffree.ffmpeg.UrlInput;
-import com.github.kokorin.jaffree.ffmpeg.UrlOutput;
 import lombok.extern.slf4j.Slf4j;
-
-import javax.imageio.ImageIO;
-import java.awt.image.BufferedImage;
-import java.nio.file.Paths;
 
 /**
  * 图片工具类
@@ -23,24 +16,47 @@ public class ImageUtils {
      * TODO svg、webp格式文件缩略图待完善
      */
     public static Boolean generateThumbnail(String srcImagePath, String destImagePath, int width, int height) {
+        // 获取宽度命令
+//        String getWidthCommand = String.format("identify -format \"%%w\" %s", srcImagePath);
+        // ImageMagick生成缩略图命令
+        String command = String.format("convert %s -resize 150x -quality 80 %s", srcImagePath, destImagePath);
+
         try {
-            BufferedImage src = ImageIO.read(FileUtil.file(srcImagePath));
-            if (src != null) {
-                int sorceW = src.getWidth();
-                int sorceH = src.getHeight();
-                // 小于 指定宽度不压缩
-                if (sorceW <= width) {
-                    FileUtil.copy(srcImagePath, destImagePath, true);
-                    return true;
-                }
-            }
+            // 执行命令并获取图片宽度
+//            String widthStr = RuntimeUtil.execForStr(getWidthCommand);
+//            Integer imgWidth = null;
+//
+//            try {
+//                // 转换为整数
+//                imgWidth = NumberUtil.parseInt(StrUtil.removeAll(widthStr, '"').trim());
+//            } catch (NumberFormatException e) {
+//                log.error("获取图片宽度失败，无法解析宽度值，宽度返回结果为：" + widthStr);
+//            }
+//
+//            // 小于指定宽度不压缩
+//            if (imgWidth != null && imgWidth <= width) {
+//                FileUtil.copy(srcImagePath, destImagePath, true);
+//                return true;
+//            }
 
             if (!FileUtil.exist(destImagePath)) {
                 FileUtil.touch(destImagePath);
             }
 
-            FFmpeg.atPath().addInput(UrlInput.fromPath(Paths.get(srcImagePath))).setOverwriteOutput(true).addArguments("-vf", "scale=" + width + ":-1").addArguments("-q:v", "2") // 这里"-1"表示，高度将会自动按照输入视频的长宽比进行调整
-                    .addOutput(UrlOutput.toPath(Paths.get(destImagePath))).execute();
+            // 执行命令
+            Process process = RuntimeUtil.exec(command);
+
+            // 等待命令执行完成
+            int exitCode = process.waitFor();
+            if (exitCode != 0) {
+                String errorOutput = RuntimeUtil.getErrorResult(process);
+                log.error("生成缩略图失败，失败原因：" + errorOutput);
+                FileUtil.del(destImagePath);
+                return false;
+            }
+
+//            FFmpeg.atPath().addInput(UrlInput.fromPath(Paths.get(srcImagePath))).setOverwriteOutput(true).addArguments("-vf", "scale=" + width + ":-1").addArguments("-q:v", "2") // 这里"-1"表示，高度将会自动按照输入视频的长宽比进行调整
+//                    .addOutput(UrlOutput.toPath(Paths.get(destImagePath))).execute();
         } catch (Exception e) {
             log.error("生成缩略图失败，失败原因：{}", e.getMessage());
             FileUtil.del(destImagePath);