+++ aether-core/src/main/java/com/xiaobai1226/aether/core/controller/UserController.java
@@ -2,6 +2,7 @@ package com.xiaobai1226.aether.core.controller;
 
 import cn.dev33.satoken.stp.StpUtil;
 import cn.hutool.core.io.FileUtil;
+import cn.hutool.core.io.resource.ResourceUtil;
 import com.xiaobai1226.aether.common.constant.FolderNameConsts;
 import com.xiaobai1226.aether.common.constant.SystemConsts;
 import com.xiaobai1226.aether.core.domain.dto.UserSpaceUsageDTO;
@@ -15,14 +16,14 @@ import org.noear.solon.annotation.*;
 import org.noear.solon.core.handle.Context;
 import org.noear.solon.core.handle.DownloadedFile;
 import org.noear.solon.core.handle.UploadedFile;
-import org.noear.solon.core.util.ResourceUtil;
 import org.noear.solon.validation.annotation.Valid;
 import org.noear.solon.validation.annotation.Validated;
 
 import java.io.IOException;
-import java.net.URL;
+import java.io.InputStream;
 import java.nio.file.Path;
 
+import static cn.hutool.http.ContentType.OCTET_STREAM;
 import static com.xiaobai1226.aether.common.constant.GateWayTagConsts.API_V1;
 import static com.xiaobai1226.aether.common.constant.ResultErrorMsgConsts.ERROR_UPDATE_PASSWORD;
 import static com.xiaobai1226.aether.common.enums.ResultCodeEnum.BAD_REQUEST_ERROR;
@@ -86,28 +87,23 @@ public class UserController {
 
         // 设置头像存储全路径
         var avatarPath = FileUtils.generatePath(avatarFolderPath, userId + SystemConsts.AVATAR_SUFFIX);
-//        ctx.contentType("image/jpg");
-        // 判断文件是否存在
-//        if (FileUtil.exist(avatarPath)) {
-//            FileUtils.readFile(ctx, avatarPath);
-//        } else {
-//            var defaultAvatarPath = FileUtils.generatePath("/avatar", SystemConsts.DEFAULT_AVATAR_FILE_NAME);
-//            FileUtils.readFileByResources(ctx, defaultAvatarPath);
-//        }
 
         try {
-            URL res = null;
+            InputStream inputStream = null;
 
             // 判断文件是否存在
             if (!FileUtil.exist(avatarPath)) {
                 avatarPath = FileUtils.generatePath("avatar", SystemConsts.DEFAULT_AVATAR_FILE_NAME);
-//                res = UserController.class.getClassLoader().getResource(avatarPath);
-                res = ResourceUtil.getResourceByFile(avatarPath);
+                inputStream = ResourceUtil.getStream(avatarPath);
             }
 
-            var file = (res == null) ? FileUtil.file(avatarPath) : FileUtil.file(res);
+            DownloadedFile downloadedFile;
 
-            var downloadedFile = new DownloadedFile(file);
+            if (inputStream == null) {
+                downloadedFile = new DownloadedFile(FileUtil.file(avatarPath));
+            } else {
+                downloadedFile = new DownloadedFile(OCTET_STREAM.getValue(), inputStream, "avatar.png");
+            }
 
             // 不做为附件下载（按需配置）
             downloadedFile.asAttachment(false);